Create listComprehension.py
You are given three integers  x,y and z  representing the dimensions of a cuboid along with an integer n. Print a list of all possible coordinates given by (i,j,k)on a 3D grid where the sum of i+j+k is not equal to n. Example
Input 
1
1
1
2

Output
[[0, 0, 0], [0, 0, 1], [0, 1, 0], [1, 0, 0], [1, 1, 1]]