Create AverageMarks.py
The first line contains the integer n, the number of students' records. The next n lines contain the names and marks obtained by a student, each value separated by a space. The final line contains query_name, the name of a student to query.
Print one line: The average of the marks obtained by the particular student correct to 2 decimal places.