Create switchGame.py
Chef has N switches which are currently OFF. Chef is sleeping. There are N kids denoted by ki (1≤i≤N) i.e. k1, k2 and so on. ki kid comes and changes the state of switches of every ith switch. For e.g., if k1 kid comes, he turns on all the switches, and then if k2 kid comes, he turns off every second switch as all the switches are ON now.

After kN kid changes the state of every nth lock, Chef wakes up. Chef wants you to output all the positions of the switch that is currently ON.

Input:
First-line will contain T, the number of test cases. Then the test cases follow.
Each test case contains a single integer N.
Output:
For each test case, output in a single line the position of the switch that is currently ON.

Sample Input:
1
2
Sample Output:
1

EXPLANATION:
There are 2 switches. The first kid turns on both switches and the second kid turns off the second switch. So, now the first switch is ON.